cahier_de_texte/requirements.txt
flask
flask-login
flask-wtf
flask-sqlalchemy
flask-mail
email-validator
python-dotenv
openpyxl
gunicorn
Flask-Migrate
